+++ src/screens/Profile/index.js
@@ -1,7 +1,6 @@
 import EditProfileScreen from './EditProfileScreen';
 import ProfileHomeScreen from './ProfileHomeScreen';
 
-
 export {
   ProfileHomeScreen, EditProfileScreen,
 };
